srnd/src/main/java/org/trng/srnd/SrndEndpoint.java
/**
 * Licensed to the Apache Software Foundation (ASF) under one or more
 * contributor license agreements.  See the NOTICE file distributed with
 * this work for additional information regarding copyright ownership.
 * The ASF licenses this file to You under the Apache License, Version 2.0
 * (the "License"); you may not use this file except in compliance with
 * the License.  You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package org.trng.srnd;

import org.apache.camel.Consumer;
import org.apache.camel.Processor;
import org.apache.camel.Producer;
import org.apache.camel.impl.DefaultEndpoint;

/**
 * Represents a srnd endpoint.
 */
public class SrndEndpoint extends DefaultEndpoint {
	private long initialDelay = 5000;
	private long delay = 5000;
	private int size = 1024;
	private String provider = null;
	private String algorithm = null;

	public String getProvider() {
		return provider;
	}

	public void setProvider(String provider) {
		this.provider = provider;
	}

	public SrndEndpoint() {
		super();
	}

	public SrndEndpoint(String uri, String algorithm, SrndComponent component) {
		super(uri, component);
		setAlgorithm(algorithm);
	}

	public long getDelay() {
		return delay;
	}

	public void setDelay(long delay) {
		this.delay = delay;
	}

	public int getSize() {
		return size;
	}

	public void setSize(int size) {
		this.size = size;
	}

	public String getAlgorithm() {
		return algorithm;
	}

	public void setAlgorithm(String algorithm) {
		this.algorithm = algorithm;
	}

	public Producer createProducer() throws Exception {
		return new SrndProducer(this);
	}

	public Consumer createConsumer(Processor processor) throws Exception {
		return new SrndConsumer(this, processor);
	}

	public boolean isSingleton() {
		return true;
	}

	public long getInitialDelay() {
		return initialDelay;
	}

	public void setInitialDelay(long initialDelay) {
		this.initialDelay = initialDelay;
	}

}
